Backend/requirements.txt
flask
flask-cors
gunicorn
python-dotenv
pytesseract
pdf2image
Pillow
google-generativeai